🚀 Pre-surgery commit: Working reboot system with circular charts fix and enhanced templates
- Fixed circular coordinate visualization in statistical reports
- Enhanced template with beautiful traffic light indicators
- Robust response parser for Claude JSON handling
- Statistical comparison working with 83.1% correlation results
- Ready for clean extraction surgery

+++ src/reboot/engine/signature_engine.py
@@ -101,7 +101,7 @@ def calculate_coordinates(
             total_weight += force
 
     if total_weight > 0:
-        return weighted_x / total_weight, weighted_y / total_weight
+        return float(weighted_x / total_weight), float(weighted_y / total_weight)
 
     return 0.0, 0.0
